requirements.txt
pandas
spotipy
python-dotenv
rich
shazamio